+++ tests/scripts/start.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+
+# Envisaged Redux
+# Copyright (c) 2019 Carl Colena
+#
+# SPDX-License-Identifier: Apache-2.0
+
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 while [[ $# -gt 0 ]]; do